ADD Version 12 terminada sin version 11 y version 10

+++ src/me/franciscofl12/arkanoid/Actor.java
@@ -103,7 +103,7 @@ public abstract class Actor {
 		this.velocidadDeCambioDeSprite = velocidadDeCambioDeSprite;
 	}
 
-	public int getVelocidadX() {
+	public double getVelocidadX() {
 		return velocidadX;
 	}
